Refactor EmojiGrid component: consolidate mobile and desktop versions, enhance rendering logic, and improve visibility handling
