+++ frontend/src/components/MeetingRoom/EmojiGrid/EmojiGrid.spec.tsx
@@ -1,5 +1,5 @@
 import { afterEach, beforeEach, describe, expect, it, Mock, vi } from 'vitest';
-import { render, screen } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import { ReactElement, useRef, useState } from 'react';
 import Button from '@ui/Button';
 import useMediaQuery from '@ui/useMediaQuery';
@@ -55,7 +55,9 @@ describe('EmojiGrid', () => {
     it('displays emoji grid when open', () => {
       render(<TestComponent defaultOpenEmojiGrid />);
 
-      expect(screen.getByTestId('send-emoji-button')).toBeVisible();
+      return waitFor(() => {
+        expect(screen.getByTestId('send-emoji-button')).toBeVisible();
+      });
     });
 
     it('displays nothing when closed', () => {
@@ -73,13 +75,15 @@ describe('EmojiGrid', () => {
     it('displays emoji grid when open', () => {
       render(<TestComponent defaultOpenEmojiGrid />);
 
-      expect(screen.queryByTestId('send-emoji-button')).toBeVisible();
+      return waitFor(() => {
+        expect(screen.getByTestId('send-emoji-button')).toBeVisible();
+      });
     });
 
     it('displays nothing when closed', () => {
       render(<TestComponent defaultOpenEmojiGrid={false} />);
 
-      expect(screen.queryByTestId('send-emoji-button')).not.toBeVisible();
+      expect(screen.queryByTestId('send-emoji-button')).not.toBeInTheDocument();
     });
   });
 });
