+++ frontend/src/components/MeetingRoom/EmojiGrid/EmojiGrid.tsx
@@ -1,7 +1,22 @@
-import { Dispatch, ReactElement, RefObject, SetStateAction } from 'react';
-import EmojiGridMobile from './EmojiGridMobile';
-import EmojiGridDesktop from './EmojiGridDesktop';
+import Box from '@ui/Box';
+import ClickAwayListener from '@ui/ClickAwayListener';
+import Grid from '@ui/Grid';
+import Grow from '@ui/Grow';
+import Paper from '@ui/Paper';
+import Popper from '@ui/Popper';
+import useTheme from '@ui/theme';
+import {
+  Dispatch,
+  ReactElement,
+  RefObject,
+  SetStateAction,
+  useEffect,
+  useMemo,
+  useState,
+} from 'react';
 import useIsSmallViewport from '../../../hooks/useIsSmallViewport';
+import SendEmojiButton from '../SendEmojiButton';
+import emojiMap from '../../../utils/emojis';
 
 export type EmojiGridProps = {
   isEmojiGridOpen: boolean;
@@ -29,27 +44,78 @@ const EmojiGrid = ({
 }: EmojiGridProps): ReactElement => {
   const isSmallViewport = useIsSmallViewport();
 
-  const handleClickAway = () => {
-    if (isSmallViewport) {
+  const theme = useTheme();
+  const [isAnchorReady, setIsAnchorReady] = useState<boolean>(false);
+
+  useEffect(() => {
+    if (!anchorRef.current) {
       return;
     }
-    // If a user clicks the toggle button on mobile, we save their preference for later. If a user doesn't select
-    // the toggle, the clickAwayListener will close the grid.
+
+    setIsAnchorReady(true);
+  }, [anchorRef]);
+
+  const maxWidth = useMemo(() => {
+    if (isSmallViewport) {
+      return 'calc(100dvw - 30px)';
+    }
+
+    // Each button is 66px, 8px left and right padding = 280px
+    return '280px';
+  }, [isSmallViewport]);
+
+  const handleClickAway = () => {
     setIsEmojiGridOpen(false);
   };
 
-  return isSmallViewport ? (
-    <EmojiGridMobile
-      handleClickAway={handleClickAway}
-      isEmojiGridOpen={isEmojiGridOpen}
-      isToolbarOpen={isParentOpen}
-    />
-  ) : (
-    <EmojiGridDesktop
-      handleClickAway={handleClickAway}
-      isEmojiGridOpen={isEmojiGridOpen}
-      anchorRef={anchorRef}
-    />
+  const isOpen = isParentOpen && isEmojiGridOpen && isAnchorReady;
+
+  return (
+    <Popper open={isOpen} anchorEl={anchorRef.current} transition disablePortal placement="bottom">
+      {({ TransitionProps, placement }) => (
+        <Grow
+          {...TransitionProps}
+          style={{
+            transformOrigin: placement === 'bottom' ? 'center top' : 'center bottom',
+          }}
+        >
+          <Box
+            sx={{
+              display: 'flex',
+              textAlign: 'left',
+              fontWeight: 'normal',
+            }}
+          >
+            <ClickAwayListener onClickAway={handleClickAway}>
+              <Paper
+                data-testid="emoji-grid"
+                sx={{
+                  display: 'flex',
+                  alignItems: 'center',
+                  justifyContent: 'center',
+                  backgroundColor: theme.colors.darkGreyOpacity,
+                  color: theme.colors.onDarkGrey,
+                  padding: { xs: 1 },
+                  borderRadius: 2,
+                  zIndex: 1,
+                  transform: 'translateY(-5%)',
+                  width: maxWidth,
+                  maxWidth,
+                  mx: 2,
+                  position: 'relative',
+                }}
+              >
+                <Grid container spacing={0} sx={{ width: '100%' }}>
+                  {Object.values(emojiMap).map((emoji) => (
+                    <SendEmojiButton key={emoji} emoji={emoji} />
+                  ))}
+                </Grid>
+              </Paper>
+            </ClickAwayListener>
+          </Box>
+        </Grow>
+      )}
+    </Popper>
   );
 };
