--- frontend/src/components/MeetingRoom/EmojiGrid/EmojiGridDesktop.tsx
@@ -1,103 +1,0 @@
-import Grid from '@ui/Grid';
-import Grow from '@ui/Grow';
-import Paper from '@ui/Paper';
-import Popper from '@ui/Popper';
-import ClickAwayListener from '@ui/ClickAwayListener';
-import Box from '@ui/Box';
-import { ReactElement, RefObject, useEffect, useState } from 'react';
-import useTheme from '@ui/theme';
-import SendEmojiButton from '../SendEmojiButton';
-import emojiMap from '../../../utils/emojis';
-
-export type EmojiGridDesktopProps = {
-  handleClickAway: (event: MouseEvent | TouchEvent) => void;
-  isEmojiGridOpen: boolean;
-  anchorRef: RefObject<HTMLButtonElement | null>;
-};
-
-/**
- * EmojiGridDesktop Component
- *
- * Displays a grid of emojis for devices with larger viewports.
- * @param {EmojiGridDesktopProps} props - the props for the component
- *  @property {(event: MouseEvent | TouchEvent) => void} handleClickAway - handles clicking away from the emoji grid
- *  @property {boolean} isEmojiGridOpen - whether the component is open
- *  @property {RefObject<HTMLButtonElement | null>} anchorRef - the button ref for the grid
- * @returns {ReactElement | false} - The EmojiGridDesktop Component
- */
-
-const EmojiGridDesktop = ({
-  handleClickAway,
-  isEmojiGridOpen,
-  anchorRef,
-}: EmojiGridDesktopProps): ReactElement | false => {
-  const theme = useTheme();
-  const [isRendered, setIsRendered] = useState<boolean>(false);
-  useEffect(() => {
-    // useRef is not immediately assigned on first render
-    // We ensure the component renders only after the anchorRef is assigned
-    if (anchorRef.current) {
-      setIsRendered(true);
-    }
-  }, [anchorRef]);
-
-  return (
-    isRendered && (
-      <Popper
-        open={isEmojiGridOpen}
-        anchorEl={anchorRef.current}
-        transition
-        disablePortal
-        placement="bottom"
-      >
-        {({ TransitionProps, placement }) => (
-          <Grow
-            {...TransitionProps}
-            style={{ transformOrigin: placement === 'bottom' ? 'center top' : 'center bottom' }}
-          >
-            <Box
-              sx={{
-                display: 'flex',
-                textAlign: 'left',
-                fontWeight: 'normal',
-              }}
-            >
-              <ClickAwayListener onClickAway={handleClickAway}>
-                <Paper
-                  data-testid="emoji-grid"
-                  sx={{
-                    display: 'flex',
-                    alignItems: 'center',
-                    justifyContent: 'center',
-                    backgroundColor: theme.colors.darkGreyOpacity,
-                    color: theme.colors.onDarkGrey,
-                    padding: { xs: 1 },
-                    borderRadius: 2,
-                    zIndex: 1,
-                    transform: 'translateY(-5%)',
-                    // Each button is 66px, 8px left and right padding = 280px
-                    maxWidth: '280px',
-                    position: 'relative',
-                  }}
-                >
-                  <Grid
-                    container
-                    spacing={0}
-                    display={isEmojiGridOpen ? 'flex' : 'none'}
-                    sx={{ width: '100%' }}
-                  >
-                    {Object.values(emojiMap).map((emoji) => (
-                      <SendEmojiButton key={emoji} emoji={emoji} />
-                    ))}
-                  </Grid>
-                </Paper>
-              </ClickAwayListener>
-            </Box>
-          </Grow>
-        )}
-      </Popper>
-    )
-  );
-};
-
-export default EmojiGridDesktop;
